step#2 additional ver - 라스트 프레임 보너스 핀 저장소 추상화
